Further split views into subviews. Extract logic into viewmodel.
